patch: move code for server setup and health check to correct test-setup.ts run by viteset
